# Parallelize Resolver