--- a/crates/uv-interpreter/src/environment.rs
+++ b/crates/uv-interpreter/src/environment.rs
@@ -1,6 +1,7 @@
 use itertools::Either;
 use std::env;
 use std::path::{Path, PathBuf};
+use std::sync::Arc;
 
 use same_file::is_same_file;
 
@@ -12,7 +13,10 @@ use crate::{find_default_python, find_requested_python, Error, Interpreter, Targ
 
 /// A Python environment, consisting of a Python [`Interpreter`] and its associated paths.
 #[derive(Debug, Clone)]
-pub struct PythonEnvironment {
+pub struct PythonEnvironment(Arc<SharedPythonEnvironment>);
+
+#[derive(Debug, Clone)]
+struct SharedPythonEnvironment {
     root: PathBuf,
     interpreter: Interpreter,
 }
@@ -46,10 +50,10 @@ impl PythonEnvironment {
             interpreter.base_prefix().display()
         );
 
-        Ok(Self {
+        Ok(Self(Arc::new(SharedPythonEnvironment {
             root: venv,
             interpreter,
-        })
+        })))
     }
 
     /// Create a [`PythonEnvironment`] for a Python interpreter specifier (e.g., a path or a binary name).
@@ -57,57 +61,58 @@ impl PythonEnvironment {
         let Some(interpreter) = find_requested_python(python, cache)? else {
             return Err(Error::RequestedPythonNotFound(python.to_string()));
         };
-        Ok(Self {
+        Ok(Self(Arc::new(SharedPythonEnvironment {
             root: interpreter.prefix().to_path_buf(),
             interpreter,
-        })
+        })))
     }
 
     /// Create a [`PythonEnvironment`] for the default Python interpreter.
     pub fn from_default_python(cache: &Cache) -> Result<Self, Error> {
         let interpreter = find_default_python(cache)?;
-        Ok(Self {
+        Ok(Self(Arc::new(SharedPythonEnvironment {
             root: interpreter.prefix().to_path_buf(),
             interpreter,
-        })
+        })))
     }
 
     /// Create a [`PythonEnvironment`] from an existing [`Interpreter`] and root directory.
     pub fn from_interpreter(interpreter: Interpreter) -> Self {
-        Self {
+        Self(Arc::new(SharedPythonEnvironment {
             root: interpreter.prefix().to_path_buf(),
             interpreter,
-        }
+        }))
     }
 
     /// Create a [`PythonEnvironment`] from an existing [`Interpreter`] and `--target` directory.
     #[must_use]
     pub fn with_target(self, target: Target) -> Self {
-        Self {
-            interpreter: self.interpreter.with_target(target),
-            ..self
-        }
+        let inner = Arc::unwrap_or_clone(self.0);
+        Self(Arc::new(SharedPythonEnvironment {
+            interpreter: inner.interpreter.with_target(target),
+            ..inner
+        }))
     }
 
     /// Returns the root (i.e., `prefix`) of the Python interpreter.
     pub fn root(&self) -> &Path {
-        &self.root
+        &self.0.root
     }
 
     /// Return the [`Interpreter`] for this virtual environment.
     pub fn interpreter(&self) -> &Interpreter {
-        &self.interpreter
+        &self.0.interpreter
     }
 
     /// Return the [`PyVenvConfiguration`] for this virtual environment, as extracted from the
     /// `pyvenv.cfg` file.
     pub fn cfg(&self) -> Result<PyVenvConfiguration, Error> {
-        Ok(PyVenvConfiguration::parse(self.root.join("pyvenv.cfg"))?)
+        Ok(PyVenvConfiguration::parse(self.0.root.join("pyvenv.cfg"))?)
     }
 
     /// Returns the location of the Python executable.
     pub fn python_executable(&self) -> &Path {
-        self.interpreter.sys_executable()
+        self.0.interpreter.sys_executable()
     }
 
     /// Returns an iterator over the `site-packages` directories inside a virtual environment.
@@ -118,11 +123,11 @@ impl PythonEnvironment {
     /// Some distributions also create symbolic links from `purelib` to `platlib`; in such cases, we
     /// still deduplicate the entries, returning a single path.
     pub fn site_packages(&self) -> impl Iterator<Item = &Path> {
-        if let Some(target) = self.interpreter.target() {
+        if let Some(target) = self.0.interpreter.target() {
             Either::Left(std::iter::once(target.root()))
         } else {
-            let purelib = self.interpreter.purelib();
-            let platlib = self.interpreter.platlib();
+            let purelib = self.0.interpreter.purelib();
+            let platlib = self.0.interpreter.platlib();
             Either::Right(std::iter::once(purelib).chain(
                 if purelib == platlib || is_same_file(purelib, platlib).unwrap_or(false) {
                     None
@@ -135,31 +140,31 @@ impl PythonEnvironment {
 
     /// Returns the path to the `bin` directory inside a virtual environment.
     pub fn scripts(&self) -> &Path {
-        self.interpreter.scripts()
+        self.0.interpreter.scripts()
     }
 
     /// Grab a file lock for the virtual environment to prevent concurrent writes across processes.
     pub fn lock(&self) -> Result<LockedFile, std::io::Error> {
-        if let Some(target) = self.interpreter.target() {
+        if let Some(target) = self.0.interpreter.target() {
             // If we're installing into a `--target`, use a target-specific lock file.
             LockedFile::acquire(
                 target.root().join(".lock"),
                 target.root().simplified_display(),
             )
-        } else if self.interpreter.is_virtualenv() {
+        } else if self.0.interpreter.is_virtualenv() {
             // If the environment a virtualenv, use a virtualenv-specific lock file.
-            LockedFile::acquire(self.root.join(".lock"), self.root.simplified_display())
+            LockedFile::acquire(self.0.root.join(".lock"), self.0.root.simplified_display())
         } else {
             // Otherwise, use a global lock file.
             LockedFile::acquire(
-                env::temp_dir().join(format!("uv-{}.lock", cache_key::digest(&self.root))),
-                self.root.simplified_display(),
+                env::temp_dir().join(format!("uv-{}.lock", cache_key::digest(&self.0.root))),
+                self.0.root.simplified_display(),
             )
         }
     }
 
     /// Return the [`Interpreter`] for this virtual environment.
     pub fn into_interpreter(self) -> Interpreter {
-        self.interpreter
+        Arc::unwrap_or_clone(self.0).interpreter
     }
 }