--- a/crates/uv-requirements/src/unnamed.rs
+++ b/crates/uv-requirements/src/unnamed.rs
@@ -1,6 +1,7 @@
 use std::borrow::Cow;
 use std::path::Path;
 use std::str::FromStr;
+use std::sync::Arc;
 
 use anyhow::Result;
 use configparser::ini::Ini;
@@ -254,13 +255,18 @@ impl<'a, Context: BuildContext> NamedRequirementsResolver<'a, Context> {
         // Fetch the metadata for the distribution.
         let name = {
             let id = VersionId::from_url(source.url());
-            if let Some(archive) = index.get_metadata(&id).as_deref().and_then(|response| {
-                if let MetadataResponse::Found(archive) = response {
-                    Some(archive)
-                } else {
-                    None
-                }
-            }) {
+            if let Some(archive) = index
+                .distributions()
+                .get(&id)
+                .as_deref()
+                .and_then(|response| {
+                    if let MetadataResponse::Found(archive) = response {
+                        Some(archive)
+                    } else {
+                        None
+                    }
+                })
+            {
                 // If the metadata is already in the index, return it.
                 archive.metadata.name.clone()
             } else {
@@ -272,7 +278,9 @@ impl<'a, Context: BuildContext> NamedRequirementsResolver<'a, Context> {
                 let name = archive.metadata.name.clone();
 
                 // Insert the metadata into the index.
-                index.insert_metadata(id, MetadataResponse::Found(archive));
+                index
+                    .distributions()
+                    .done(id, Arc::new(MetadataResponse::Found(archive)));
 
                 name
             }