--- a/crates/uv-resolver/src/error.rs
+++ b/crates/uv-resolver/src/error.rs
@@ -1,7 +1,6 @@
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::Formatter;
 use std::ops::Deref;
-use std::rc::Rc;
 use std::sync::Arc;
 
 use indexmap::IndexMap;
@@ -239,7 +238,7 @@ impl NoSolutionError {
         mut self,
         python_requirement: &PythonRequirement,
         visited: &SharedSet<PackageName>,
-        package_versions: &OnceMap<PackageName, Rc<VersionsResponse>>,
+        package_versions: &OnceMap<PackageName, Arc<VersionsResponse>>,
     ) -> Self {
         let mut available_versions = IndexMap::default();
         for package in self.derivation_tree.packages() {
@@ -263,7 +262,7 @@ impl NoSolutionError {
                     // tree, but were never visited during resolution. We _may_ have metadata for
                     // these packages, but it's non-deterministic, and omitting them ensures that
                     // we represent the state of the resolver at the time of failure.
-                    if visited.borrow().contains(name) {
+                    if visited.contains(name) {
                         if let Some(response) = package_versions.get(name) {
                             if let VersionsResponse::Found(ref version_maps) = *response {
                                 for version_map in version_maps {
@@ -304,7 +303,6 @@ impl NoSolutionError {
         mut self,
         unavailable_packages: &SharedMap<PackageName, UnavailablePackage>,
     ) -> Self {
-        let unavailable_packages = unavailable_packages.borrow();
         let mut new = FxHashMap::default();
         for package in self.derivation_tree.packages() {
             if let PubGrubPackage::Package(name, _, _) = package {
@@ -324,11 +322,11 @@ impl NoSolutionError {
         incomplete_packages: &SharedMap<PackageName, SharedMap<Version, IncompletePackage>>,
     ) -> Self {
         let mut new = FxHashMap::default();
-        let incomplete_packages = incomplete_packages.borrow();
         for package in self.derivation_tree.packages() {
             if let PubGrubPackage::Package(name, _, _) = package {
                 if let Some(versions) = incomplete_packages.get(name) {
-                    for (version, reason) in versions.borrow().iter() {
+                    for entry in versions.iter() {
+                        let (version, reason) = entry.pair();
                         new.entry(name.clone())
                             .or_insert_with(BTreeMap::default)
                             .insert(version.clone(), reason.clone());