… Cargo.lock
… crates/bench/benches/uv.rs
… crates/once-map/Cargo.toml
… crates/uv-resolver/Cargo.toml
… crates/uv-resolver/src/resolver/index.rs
… crates/uv-resolver/src/resolver/mod.rs
… crates/uv-resolver/tests/resolver.rs
… crates/uv-types/src/traits.rs
… crates/uv/src/commands/pip/compile.rs
… crates/uv/src/commands/pip/editables.rs
… crates/uv/src/commands/pip/install.rs
… crates/uv/src/commands/pip/sync.rs
… crates/uv/src/commands/project/lock.rs
… crates/uv/src/commands/project/mod.rs
… crates/uv/src/commands/project/run.rs