@konstin 2024-05-17
Amazing work!

Do you know why it gets so much faster, i.e. how the solver is blocking? I've been looking at the spans, but i don't really understand why the prefetches don't get queued anyway on main.

Airflow, main vs. PR:

![main](https://github.com/astral-sh/uv/assets/6826232/b7c2de41-d4e8-45e1-b323-b2a23e053b09)

![PR](https://github.com/astral-sh/uv/assets/6826232/21ab014e-bfdb-4b52-89e4-5171ddcca6f7)



---

Here's some perf number from my machine:

```
jupyter:
  Time (mean ± σ):      15.9 ms ±   1.3 ms    [User: 14.1 ms, System: 21.0 ms]
  Time (mean ± σ):      15.4 ms ±   1.1 ms    [User: 14.1 ms, System: 20.1 ms]
boto3:
  Time (mean ± σ):     383.5 ms ±   3.3 ms    [User: 343.5 ms, System: 60.8 ms]
  Time (mean ± σ):     325.3 ms ±   3.8 ms    [User: 351.7 ms, System: 62.0 ms]
airflow:
  Time (mean ± σ):     192.6 ms ±   3.9 ms    [User: 172.5 ms, System: 151.3 ms]
  Time (mean ± σ):     147.0 ms ±   3.3 ms    [User: 168.4 ms, System: 146.6 ms]
```